+++ test/kubernetes/e2e/features/tracing/testdata/otlpeek.yaml
@@ -1,0 +1,123 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: otlpeek-config
+data:
+  config.yaml: |
+    receivers:
+      otlp:
+        protocols:
+          grpc:
+            endpoint: 0.0.0.0:4317
+          http:
+            endpoint: 0.0.0.0:4318
+
+    processors:
+      batch:
+        timeout: 1s
+        send_batch_size: 1024
+
+    exporters:
+      otlpeek:
+        endpoint: ":8080"
+
+    service:
+      pipelines:
+        traces:
+          receivers: [otlp]
+          processors: [batch]
+          exporters: [otlpeek]
+        metrics:
+          receivers: [otlp]
+          processors: [batch]
+          exporters: [otlpeek]
+        logs:
+          receivers: [otlp]
+          processors: [batch]
+          exporters: [otlpeek]
+
+      telemetry:
+        logs:
+          level: info
+        metrics:
+          level: none
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: otlpeek
+  labels:
+    app: otlpeek
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: otlpeek
+  template:
+    metadata:
+      labels:
+        app: otlpeek
+    spec:
+      containers:
+      - name: otlpeek
+        image: fktkrt/otlpeek:0.0.2
+        imagePullPolicy: Always
+        ports:
+        - containerPort: 8080
+          name: web
+        - containerPort: 4317
+          name: otlp-grpc
+        - containerPort: 4318
+          name: otlp-http
+        volumeMounts:
+        - name: config
+          mountPath: /etc/otlpeek
+        resources:
+          requests:
+            memory: "64Mi"
+            cpu: "50m"
+          limits:
+            memory: "256Mi"
+            cpu: "500m"
+        livenessProbe:
+          httpGet:
+            path: /
+            port: 8080
+          initialDelaySeconds: 30
+          periodSeconds: 10
+        readinessProbe:
+          httpGet:
+            path: /
+            port: 8080
+          initialDelaySeconds: 5
+          periodSeconds: 5
+      volumes:
+      - name: config
+        configMap:
+          name: otlpeek-config
+
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: otlpeek
+  labels:
+    app: otlpeek
+spec:
+  type: ClusterIP
+  ports:
+  - port: 8080
+    targetPort: 8080
+    name: web
+    protocol: TCP
+  - port: 4317
+    targetPort: 4317
+    name: otlp-grpc
+    protocol: TCP
+    appProtocol: grpc
+  - port: 4318
+    targetPort: 4318
+    name: otlp-http
+    protocol: TCP
+  selector:
+    app: otlpeek
