+++ test/kubernetes/e2e/features/tracing/testdata/tracing-policy.yaml
@@ -1,0 +1,18 @@
+apiVersion: gateway.kgateway.dev/v1alpha1
+kind: HTTPListenerPolicy
+metadata:
+  name: tracing-policy
+spec:
+  targetRefs:
+  - group: gateway.networking.k8s.io
+    kind: Gateway
+    name: gw
+  tracing:
+    provider:
+      openTelemetry:
+        serviceName: "my:service"
+        grpcService:
+          backendRef:
+            name: otlpeek
+            namespace: default
+            port: 4317
