+++ test/kubernetes/e2e/features/tracing/types.go
@@ -1,42 +1,73 @@
-//go:build ignore
-
 package tracing
 
 import (
 	"path/filepath"
 
+	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	"github.com/kgateway-dev/kgateway/v2/api/v1alpha1"
 	"github.com/kgateway-dev/kgateway/v2/pkg/utils/fsutils"
-)
-
-const (
-	pathWithRouteDescriptor    = "/path/with/route/descriptor"
-	pathWithoutRouteDescriptor = "/path/without/route/descriptor"
-	routeDescriptorSpanName    = "THISISAROUTEDESCRIPTOR"
-	gatewayProxyHost           = "gateway-proxy-tracing"
-	gatewayProxyPort           = 18080
+	e2edefaults "github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e/defaults"
+	"github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e/tests/base"
 )
 
 var (
-	setupOtelcolManifest        = filepath.Join(fsutils.MustGetThisDir(), "testdata", "setup-otelcol.yaml")
-	tracingConfigManifest       = filepath.Join(fsutils.MustGetThisDir(), "testdata", "tracing.yaml")
-	gatewayConfigManifest       = filepath.Join(fsutils.MustGetThisDir(), "testdata", "gateway.yaml")
-	gatewayProxyServiceManifest = filepath.Join(fsutils.MustGetThisDir(), "testdata", "gw-proxy-tracing-service.yaml")
+	// manifests
+	setupManifest         = filepath.Join(fsutils.MustGetThisDir(), "testdata", "setup.yaml")
+	otelCollectorManifest = filepath.Join(fsutils.MustGetThisDir(), "testdata", "otlpeek.yaml")
+	policyManifest        = filepath.Join(fsutils.MustGetThisDir(), "testdata", "tracing-policy.yaml")
 
-	otelcolPod = &corev1.Pod{
-		ObjectMeta: metav1.ObjectMeta{Name: "otel-collector", Namespace: "default"},
-	}
-	otelcolSelector = metav1.ListOptions{
-		LabelSelector: "app.kubernetes.io/name=otel-collector",
-	}
-	otelcolUpstream = &metav1.ObjectMeta{
-		Name:      "opentelemetry-collector",
+	// setup objects
+	proxyObjectMeta = metav1.ObjectMeta{
+		Name:      "gw",
 		Namespace: "default",
 	}
-	tracingVs = &metav1.ObjectMeta{
-		Name:      "virtual-service",
-		Namespace: "default",
+	proxyDeployment = &appsv1.Deployment{ObjectMeta: proxyObjectMeta}
+	proxyService    = &corev1.Service{ObjectMeta: proxyObjectMeta}
+
+	httpbinSvc = &corev1.Service{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "httpbin",
+			Namespace: "httpbin",
+		},
+	}
+	httpbinDeployment = &appsv1.Deployment{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "httpbin",
+			Namespace: "httpbin",
+		},
+	}
+
+	// otelCollector objects
+	otelCollectorDeployment = &appsv1.Deployment{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "otlpeek",
+			Namespace: "default",
+		},
+	}
+
+	tracingPolicy = &v1alpha1.HTTPListenerPolicy{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "tracing-policy",
+			Namespace: "default",
+		},
+	}
+
+	setup = base.SimpleTestCase{
+		Manifests: []string{e2edefaults.CurlPodManifest, setupManifest},
+		Resources: []client.Object{e2edefaults.CurlPod, proxyDeployment, proxyService, httpbinSvc, httpbinDeployment},
+	}
+
+	// test cases
+	testCases = map[string]*base.TestCase{
+		"TestOTelTracing": {
+			SimpleTestCase: base.SimpleTestCase{
+				Manifests: []string{otelCollectorManifest, policyManifest},
+				Resources: []client.Object{otelCollectorDeployment},
+			},
+		},
 	}
 )
