+++ test/kubernetes/e2e/features/tracing_legacy/testdata/gateway.yaml
