+++ test/kubernetes/e2e/features/tracing_legacy/testdata/gw-proxy-tracing-service.yaml
