+++ test/kubernetes/e2e/tests/tracing_tests.go
@@ -1,0 +1,12 @@
+package tests
+
+import (
+	"github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e"
+	"github.com/kgateway-dev/kgateway/v2/test/kubernetes/e2e/features/tracing"
+)
+
+func TracingSuiteRunner() e2e.SuiteRunner {
+	suiteRunner := e2e.NewSuiteRunner(false)
+	suiteRunner.Register("Tracing", tracing.NewTestingSuite)
+	return suiteRunner
+}
